+++ packages/iov-crypto/types/index.d.ts
@@ -1,25 +1,22 @@
-export { Bip39 } from "./bip39";
-export { EnglishMnemonic } from "./englishmnemonic";
-export { HashFunction } from "./hash";
-export { Hmac } from "./hmac";
-export { Keccak256 } from "./keccak";
 export {
-  Xchacha20poly1305Ietf,
-  Xchacha20poly1305IetfCiphertext,
-  Xchacha20poly1305IetfKey,
-  Xchacha20poly1305IetfMessage,
-  Xchacha20poly1305IetfNonce,
+  Bip39,
+  EnglishMnemonic,
+  HashFunction,
+  Hmac,
+  Keccak256,
   Argon2id,
   Argon2idOptions,
   Ed25519,
   Ed25519Keypair,
-} from "./libsodium";
-export { Random } from "./random";
-export { Ripemd160 } from "./ripemd";
-export { Secp256k1, Secp256k1Keypair } from "./secp256k1";
-export { ExtendedSecp256k1Signature, Secp256k1Signature } from "./secp256k1signature";
-export { Sha1, Sha256, Sha512 } from "./sha";
-export {
+  Random,
+  Ripemd160,
+  Secp256k1,
+  Secp256k1Keypair,
+  ExtendedSecp256k1Signature,
+  Secp256k1Signature,
+  Sha1,
+  Sha256,
+  Sha512,
   pathToString,
   stringToPath,
   Slip10,
@@ -27,4 +24,9 @@ export {
   Slip10RawIndex,
   Slip10Result,
   slip10CurveFromString,
-} from "./slip10";
+  Xchacha20poly1305Ietf,
+  Xchacha20poly1305IetfCiphertext,
+  Xchacha20poly1305IetfKey,
+  Xchacha20poly1305IetfMessage,
+  Xchacha20poly1305IetfNonce,
+} from "@cosmjs/crypto";